+++ library/ctlibstructureaddon/ct_itemdrawable/ct_grid3d_sparse.h
@@ -84,9 +84,9 @@ public:
     CT_Grid3D_Sparse(double xmin,
                      double ymin,
                      double zmin,
-                     size_t dimx,
-                     size_t dimy,
-                     size_t dimz,
+                     int dimx,
+                     int dimy,
+                     int dimz,
                      double resolution,
                      DataT na,
                      DataT initValue);
@@ -195,7 +195,7 @@ public:
       * \param value Value
       * \return True if the value has actually been set
       */
-    bool setValue(const size_t colx, const size_t liny, const size_t levz, DataT value);
+    bool setValue(const int colx, const int liny, const int levz, DataT value);
 
     /**
       * \brief Set the value for the cell containing (x,y,z) coordinate to value
@@ -251,7 +251,7 @@ public:
       * \param levz z level, first one is 0
       * \return Value at row liny and column clox
       */
-    DataT value(const size_t colx, const size_t liny, const size_t levz) const;
+    DataT value(const int colx, const int liny, const int levz) const;
 
 
     /**
@@ -328,7 +328,7 @@ public:
       * \param centermode : How should be treated the central cell ?
       * \return List of neighbours values
       */
-    QList<DataT> neighboursValues(const size_t colx, const size_t liny, const size_t levz, const size_t distance, const bool keepNAs = false, const CenterMode centermode = CT_Grid3D_Sparse::CM_DropCenter) const;
+    QList<DataT> neighboursValues(const int colx, const int liny, const int levz, const int distance, const bool keepNAs = false, const CenterMode centermode = CT_Grid3D_Sparse::CM_DropCenter) const;
 
     void getIndicesWithData(QList<size_t> &list) const;
 
@@ -353,7 +353,7 @@ public:
     CT_ITEM_COPY_IMP(CT_Grid3D_Sparse<DataT>)
 
     protected:
-        DataT       _NAdata;            /*!< Valeur codant NA */
+    DataT       _NAdata;            /*!< Valeur codant NA */
     DataT       _initData;            /*!< Valeur par defaut */
 
     DataT       _dataMax;           /*!< valeur maximale du grid*/
@@ -404,7 +404,7 @@ template<>
 CTLIBSTRUCTUREADDON_EXPORT bool CT_Grid3D_Sparse<bool>::addValueAtIndex(const size_t index, const bool value);
 
 template<>
-CTLIBSTRUCTUREADDON_EXPORT QList<bool> CT_Grid3D_Sparse<bool>::neighboursValues(const size_t colx, const size_t liny, const size_t levz, const size_t distance, const bool keepNAs, const CenterMode centermode) const;
+CTLIBSTRUCTUREADDON_EXPORT QList<bool> CT_Grid3D_Sparse<bool>::neighboursValues(const int colx, const int liny, const int levz, const int distance, const bool keepNAs, const CenterMode centermode) const;
 
 /*template<>
 CTLIBSTRUCTUREADDON_EXPORT QString CT_Grid3D_Sparse<unsigned long>::valueAtIndexAsString(const size_t index) const;
